// frontend/src/client/gen/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: "http://localhost:8080" | (string & {});
};

export type ErrorDetail = {
  /**
   * Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
   */
  location?: string;
  /**
   * Error message text
   */
  message?: string;
  /**
   * The value at the given location
   */
  value?: unknown;
};

export type ErrorModel = {
  /**
   * A URL to the JSON Schema for this object.
   */
  readonly $schema?: string;
  /**
   * A human-readable explanation specific to this occurrence of the problem.
   */
  detail?: string;
  /**
   * Optional list of individual error details
   */
  errors?: Array<ErrorDetail> | null;
  /**
   * A URI reference that identifies the specific occurrence of the problem.
   */
  instance?: string;
  /**
   * HTTP status code
   */
  status?: number;
  /**
   * A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
   */
  title?: string;
  /**
   * A URI reference to human-readable documentation for the error.
   */
  type?: string;
};

export type FileItem = {
  is_dir: boolean;
  mod_time: string;
  name: string;
  path: string;
  size: number;
};

export type ListFilesResponse = {
  /**
   * A URL to the JSON Schema for this object.
   */
  readonly $schema?: string;
  current_path: string;
  items: Array<FileItem> | null;
};

export type ErrorModelWritable = {
  /**
   * A human-readable explanation specific to this occurrence of the problem.
   */
  detail?: string;
  /**
   * Optional list of individual error details
   */
  errors?: Array<ErrorDetail> | null;
  /**
   * A URI reference that identifies the specific occurrence of the problem.
   */
  instance?: string;
  /**
   * HTTP status code
   */
  status?: number;
  /**
   * A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
   */
  title?: string;
  /**
   * A URI reference to human-readable documentation for the error.
   */
  type?: string;
};

export type ListFilesResponseWritable = {
  current_path: string;
  items: Array<FileItem> | null;
};

export type GetApiListFilesData = {
  body?: never;
  path?: never;
  query?: {
    path?: string;
  };
  url: "/api/list-files";
};

export type GetApiListFilesErrors = {
  /**
   * Error
   */
  default: ErrorModel;
};

export type GetApiListFilesError = GetApiListFilesErrors[keyof GetApiListFilesErrors];

export type GetApiListFilesResponses = {
  /**
   * OK
   */
  200: ListFilesResponse;
};

export type GetApiListFilesResponse = GetApiListFilesResponses[keyof GetApiListFilesResponses];